+++ backend/cmd/artwork/bootstrap.go
@@ -47,7 +47,7 @@ func bootstrap(ctx context.Context, logger *slog.Logger, messageBus *dorky.Messa
 
 	// Add Resize Small node (width: 30)
 	resize1Config := imagegraph.NodeConfig{
-		"width": float64(30),
+		"width": 30,
 	}
 	addResize1Cmd := application.NewAddImageGraphNodeCommand(
 		graphID,
@@ -63,7 +63,7 @@ func bootstrap(ctx context.Context, logger *slog.Logger, messageBus *dorky.Messa
 
 	// Add Blur node (radius: 2)
 	blurConfig := imagegraph.NodeConfig{
-		"radius": float64(2),
+		"radius": 2,
 	}
 	addBlurCmd := application.NewAddImageGraphNodeCommand(
 		graphID,
@@ -79,7 +79,7 @@ func bootstrap(ctx context.Context, logger *slog.Logger, messageBus *dorky.Messa
 
 	// Add Resize Large node (width: 500)
 	resize2Config := imagegraph.NodeConfig{
-		"width": float64(500),
+		"width": 500,
 	}
 	addResize2Cmd := application.NewAddImageGraphNodeCommand(
 		graphID,
